rename: 패키지 구조 명료화
