+++ WearOS/app/src/main/java/com/sandy/logisync/model/Account.kt
@@ -8,13 +8,3 @@ data class Account(
     val tel: String,
     //val duty: String,
 )
-
-fun String.toAccount(): Account {
-    val jsonObject = JSONObject(this)
-    return Account(
-        id = jsonObject.getString("id"),
-        name = jsonObject.getString("name"),
-        tel = jsonObject.getString("tel"),
-        //duty = jsonObject.getString("duty")
-    )
-}
