+++ WearOS/app/src/main/java/com/sandy/logisync/wearable/di/WearableModule.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import com.google.android.gms.wearable.CapabilityClient
 import com.google.android.gms.wearable.MessageClient
 import com.google.android.gms.wearable.Wearable
-import com.sandy.logisync.wearable.MyWearableClient
+import com.sandy.logisync.wearable.message.MyWearableClient
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
