+++ WearOS/app/src/main/java/com/sandy/logisync/wearable/message/MessagePath.kt
@@ -1,4 +1,4 @@
-package com.sandy.logisync.wearable
+package com.sandy.logisync.wearable.message
 
 enum class MessagePath(val path: String) {
     TEST("/test"),
